Move config specific logic to ConfigReport
