+++ app/models/concerns/configuration_status_scoped_search.rb
@@ -3,7 +3,7 @@ module ConfigurationStatusScopedSearch
 
   module ClassMethods
     def scoped_search_status(status, options)
-      options.merge!({ :offset => Report::METRIC.index(status.to_s), :word_size => Report::BIT_NUM })
+      options.merge!({ :offset => ConfigReport::METRIC.index(status.to_s), :word_size => ConfigReport::BIT_NUM })
       scoped_search options
     end
   end
