+++ test/unit/report_test.rb
@@ -3,85 +3,6 @@ require 'test_helper'
 class ReportTest < ActiveSupport::TestCase
   def setup
     User.current = users :admin
-    @report = ConfigReport.import read_json_fixture("report-skipped.json")
-  end
-
-  test "additional features can be registered if it's uniq" do
-    original = ReportImporter.authorized_smart_proxy_features
-    ReportImporter.register_smart_proxy_feature('Chef')
-    new = ReportImporter.authorized_smart_proxy_features
-    added = new - original
-    assert_include(added, 'Chef')
-    assert_equal 1, added.size
-
-    ReportImporter.register_smart_proxy_feature('Chef')
-    new = ReportImporter.authorized_smart_proxy_features
-    added = new - original
-    assert_include(added, 'Chef')
-    assert_equal 1, added.size
-
-    ReportImporter.unregister_smart_proxy_feature('Chef')
-    new = ReportImporter.authorized_smart_proxy_features
-    added = new - original
-    assert_equal 0, added.size
-  end
-
-  test "it should true on error? if there were errors" do
-    @report.status={"applied" => 92, "restarted" => 300, "failed" => 4, "failed_restarts" => 12, "skipped" => 3, "pending" => 0}
-    assert @report.error?
-  end
-
-  test "it should not be an error if there are only skips" do
-    @report.status={"applied" => 92, "restarted" => 300, "failed" => 0, "failed_restarts" => 0, "skipped" => 3, "pending" => 0}
-    assert !@report.error?
-  end
-
-  test "it should false on error? if there were no errors" do
-    @report.status={"applied" => 92, "restarted" => 300, "failed" => 0, "failed_restarts" => 0, "skipped" => 0, "pending" => 0}
-    assert !@report.error?
-  end
-
-  test "with named scope should return our report with applied resources" do
-    @report.status={"applied" => 15, "restarted" => 0, "failed" => 0, "failed_restarts" => 0, "skipped" => 0, "pending" => 0}
-    @report.save
-    assert Report.with("applied",14).include?(@report)
-    assert !Report.with("applied", 15).include?(@report)
-  end
-
-  test "with named scope should return our report with restarted resources" do
-    @report.status={"applied" => 0, "restarted" => 5, "failed" => 0, "failed_restarts" => 0, "skipped" => 0, "pending" => 0}
-    @report.save
-    assert Report.with("restarted").include?(@report)
-  end
-
-  test "with named scope should return our report with failed resources" do
-    @report.status={"applied" => 0, "restarted" => 0, "failed" => 9, "failed_restarts" => 0, "skipped" => 0, "pending" => 0}
-    @report.save
-    assert Report.with("failed").include?(@report)
-  end
-
-  test "with named scope should return our report with failed_restarts resources" do
-    @report.status={"applied" => 0, "restarted" => 0, "failed" => 0, "failed_restarts" => 91, "skipped" => 0, "pending" => 0}
-    @report.save
-    assert Report.with("failed_restarts").include?(@report)
-  end
-
-  test "with named scope should return our report with skipped resources" do
-    @report.status={"applied" => 0, "restarted" => 0, "failed" => 0, "failed_restarts" => 0, "skipped" => 8, "pending" => 0}
-    @report.save
-    assert Report.with("skipped").include?(@report)
-  end
-
-  test "with named scope should return our report with skipped resources when other bits are also used" do
-    @report.status={"applied" => 0, "restarted" => 0, "failed" => 9, "failed_restarts" => 4, "skipped" => 8, "pending" => 3}
-    @report.save
-    assert Report.with("skipped").include?(@report)
-  end
-
-  test "with named scope should return our report with pending resources when other bits are also used" do
-    @report.status={"applied" => 0, "restarted" => 0, "failed" => 9, "failed_restarts" => 4, "skipped" => 8, "pending" => 3}
-    @report.save
-    assert Report.with("pending").include?(@report)
   end
 
   test "should expire reports created 1 week ago" do
@@ -170,26 +91,9 @@ class ReportTest < ActiveSupport::TestCase
   end
 
   describe 'Report STI' do
-    test "Report has type" do
-      assert_respond_to(Report.new, :type)
-    end
-
     test "Report has default type" do
       report = Report.new
       assert_equal('ConfigReport', report.type)
     end
-
-    test "Report has child classes" do
-      my_report = MyReport.new
-      assert_equal('MyReport', my_report.type)
-      # test that import method can be overriden
-      assert_equal(true, my_report.import)
-    end
-  end
-end
-
-class MyReport < ::Report
-  def import
-    true
   end
 end
